requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Forms and validation
WTForms==3.1.1
email-validator==2.1.0

# Security
Werkzeug==3.0.1
bcrypt==4.1.2
PyJWT==2.8.0

# API and integrations
requests==2.31.0
twilio==8.11.0
sendgrid==6.11.0

# Task queue
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.1.0

# WebSocket
python-socketio==5.10.0
eventlet==0.33.3

# Excel/CSV export
openpyxl==3.1.2
pandas==2.1.4

# Notifications
pyfcm==1.5.4

# Development
gunicorn==21.2.0